requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
chromadb
sentence-transformers
openai
PyMuPDF
pdfplumber